# swagger/swagger.yml
openapi: 3.1.0
info:
  title: Servicio de Datos - OpenAPI 3.1
  description: |-
    Servicio de datos para el sistema software agro-predict.   
    Encargado de solicitar, acumular, normalizar y almacenar en la base de datos, datos obtenidos por servicios públicos externos.  

    Devuelve los datos en formato JSON con codificación UTF-8.

    Links sobre los servicios de abastecimiento de datos meteorológicos:
    - [Servicio AEMET](https://github.com/agro-predict-tfg-2026/Aemet.git)
    - [Servicio SiAR](https://github.com/agro-predict-tfg-2026/SiAR.git)
    - [Servicio ITACyL](https://github.com/agro-predict-tfg-2026/ITACyL.git)
    - [Servicio IA Generativa](https://github.com/agro-predict-tfg-2026/mcp-ia-service.git)

  contact:
    email: amendoma@alumnos.unex.es
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:5000
  - url: http://data-service:9002
  - url: http://51.44.230.198:9002
...
paths:
  /climate/historical/provincias:
    get:
      tags:
        - Históricos
      summary: Obtiene datos climaticos históricos por provincias
      description: Consulta el proveedor externo de SiAR para obtener los datos climáticos históricos asociados a una provincia en un rango temporal determinado.
      operationId: historicalProvincial
      parameters:
        - name: provinceCode
          in: query
          description: Código de la provincia
          required: true
          schema:
            type: string
            enum: [CC, BA]
            default: "CC"
        - name: type
          in: query
          description: Tipo de registro horario que se quiere consultar
          required: true
          schema:
            type: string
            enum: [HORA, DIA, SEMANA]
            default: "DIA"
        - name: startDate
          in: query
          description: Fecha inicial de recogida de datos
          required: true
          schema:
            type: string
            format: date
            example: "2026-01-04"
        - name: endDate
          in: query
          description: Fecha final de recogida de datos
          required: true
          schema:
            type: string
            format: date
            example: "2026-01-04"
        - name: download
          in: query
          description: Si se habilita, se descarga en el equipo un fichero json con la Respuesta
          required: false
          schema:
            type: boolean
            example: false
      responses:
        '200':
          description: Datos históricos provinciales obtenidos correctamente
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/HistoricoHoraResponse'
                  - $ref: '#/components/schemas/HistoricoDiaResponse'
                  - $ref: '#/components/schemas/HistoricoSemanaResponse'
                  - $ref: '#/components/schemas/HistoricoMesResponse'
                discriminator:
                  propertyName: type
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Todos los parámetros deben estar definidos"
                status: 400
                error: "Invalid Parameters"
        '404':
          description: No se han encontrado datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "No se han encontrado datos para los parámetros especificados"
                status: 404
                error: "Data Not Found"
        '502':
          description: Error del proveedor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Error del proveedor de datos"
                status: 502
                error: "Provider Error"
  /climate/historical/estacion:
    get:
      tags:
        - Históricos
      summary: Obtiene datos climáticos históricos por estación
      description: Consulta el proveedor externo de SiAR para obtener datos climáticos asociados a una estación en un rango temporal determinado.
      operationId: historicalEstacion
      parameters:
        - name: estacionCode
          in: query
          description: Identificador único de la estacion
          required: true
          schema:
            type: string
            example: "CC01"
        - name: type
          in: query
          description: Tipo de registro horario que se quiere consultar
          required: true
          schema:
            type: string
            enum: [HORA, DIA, SEMANA, MES]
            default: "DIA"
        - name: startDate
          in: query
          description: Fecha inicial de recogida de datos
          required: true
          schema:
            type: string
            format: date
            example: "2025-01-01"
        - name: endDate
          in: query
          description: Fecha final de recogida de datos
          required: true
          schema:
            type: string
            format: date
            example: "2025-01-01"
        - name: download
          in: query
          description: Si se habilita, se descarga en el equipo un fichero json con la Respuesta
          required: false
          schema:
            type: boolean
            example: false

      responses:
        '200':
          description: Datos históricos para la estacion `estacionCode` obtenidos correctamente
          content:
            application/json:
              schema:
                type : string
                format: binary
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Todos los parámetros deben estar definidos"
                status: 400
                error: "Invalid Parameters"
        '404':
          description: No se han encontrado datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "No se han encontrado datos para los parámetros especificados"
                status: 404
                error: "Data Not Found"
        '502':
          description: Error del proveedor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Error del proveedor de datos"
                status: 502
                error: "Provider Error"
  /climate/historical/reintentar-pendientes:
    post:
      tags:
        - Históricos
      summary: Reintenta peticiones de datos históric que fallaron
      description: Obtiene peticiones fallidas cacheadas para reintentar su recopilación de datos en el caso que actualmente existieran
      operationId: reintentar_pendientes
      responses:
        '202':
          description: Reintento aplicado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaTasks'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:  
                message: "Provider Error"
                status: 500
                error: "Error del proveedor de datos"
...
  /sensores:
    get:
      tags:
        - Sensores
      summary: Obtiene mediciones almacenadas sobre el sensor indicado
      description: Dado el identificador de sensor que se introduce como parámetro de la consulta 
        devuelve al cliente las diferentes mediciones almacenadas en el sistema sobre ese sensor en un 
        periodo de tiempo también definido en los parámetros de la consulta.
      operationId: sensores
      parameters:
        - name: eui
          in: query
          required: true
          description: Identificador único del sensor que se quiere obtener sus mediciones
          schema:
            type: string
            default: "a840419b3188cc0c"
        - name: fecha_inicio 
          in: query
          required: true
          description: Fecha de inicio para la recopilación de mediciones sobre el sensor indicado
          schema:
            type: string
            format: date
            default: "2025-02-20"
        - name: fecha_fin
          in: query
          required: true
          description: Fecha de finalización para la recopilación de mediciones sobre el sensor indicado
          schema:
            type: string
            format: date
            default: "2025-02-25"
        - name: nombre_dt_agro
          in: query
          required: true
          description: Nombre de campo-valor a extraer sobre datos referentes al sensor especificado en `eui` desde DTAgro.
          schema:
            type: string
            default: "Leaf_Moisture"
        - name: nombre_predictor
          in: query
          required: true
          description: Nombre traducido de `nombre_dt_agro` al vocabulario del predictor de plagas cuando se envíen los datos obtenidos.
          schema:
            type: string
            default: "humedad_foliar"
      responses:
        '200':
          description: Datos del sensor `{eui}` recopilados correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatosSensorEui'
        '400':
          description: Error en los parametros de la consulta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
            example:
              message: "Invalid parameters"
              status: 400
              error: 'Debe indicarse todos los parámetros de la query (eui, fecha_inicio, fecha_fin)'
        '404':
          description: Error al obtener mediciones sobre los parámetros indicados en la consulta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Data Not Found"
                status: 404
                error: "No se han encontrado datos de sensores para los parámetros indicados"
        '502':
          description: Error del proveedor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                message: "Provider Error"
                status: 502
                error: "Error del proveedor de datos"
...
components:
  schemas:
    RespuestaTasks:
      type: object
      properties:
        task_id: 
          type: string
          format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        status:
          type: string
          example: "PROCESSING"
...
    DatosSensorEui:
      type: object
      properties:
        eui:
          type: array
          description: Lista de mediciones asociadas al sensor eui
          items:
            $ref: '#/components/schemas/MedicionesSensor'
    MedicionesSensor:
      type: object
      properties:
        humedad_foliar:
          type: number
          required: true
          description: Cantidad de agua libre presente en la superficie de la hoja
          example: 1.2
        temperatura_DS18B20:
          type: integer
          required: true
          description: Humedad recopilada por el sensor DS18B20
          ezample: 23
        temperatura_hojas:
          type: float
          required: true
          description: Temperatura acumulada en las hojas de una planta
          example: 12.1
        timestamp:
          type: string
          format: date-time
          required: true
          description: Fecha exacta en la que se ha realizado la medicion
          example: '2025-02-20T00:21:18.856Z'
...
    ApiError:
      type: object
      properties:
        message: 
          type: string
        status:
          type: integer
        error:
          type: string
    HistoricoHoraResponse:
     type: object
     required: [type, datos]
     properties:
      type:
        type: string
        description: Tipo de registro horario utilizado
        enum: [HORA]
      datos:
        type: array
        description: Vector de datos
        items:
          $ref: '#/components/schemas/HoraDTO'
    HistoricoDiaResponse:
      type: object
      required: [type, datos]
      properties:
        type:
          type: string
          description: Tipo de registro horario utilizado
          enum: [DIA]
        datos:
          type: array
          description: Vector de datos
          items:
            $ref: '#/components/schemas/DiaDTO'
    HistoricoSemanaResponse:
      type: object
      required: [type, datos]
      properties:
        type: 
          type: string
          description: Tipo de registro horario utilizado
          enum: [SEMANA]
        datos:
          type: array
          description: Vector de Datos
          items:
            $ref: '#/components/schemas/SemanaDTO'
    HistoricoMesResponse:
      type: object
      required: [type, datos]
      properties:
        type: 
          type: string
          description: Tipo de registro horario utilizado
          enum: [MES]
        datos:
          type: array
          description: Vector de Datos
          items:
            $ref: '#/components/schemas/MesDTO'
    HoraDTO:
      type: object
      required: [horaMin, tempMedia, humedadMedia, velViento, precipitacion, estacion, provincia, fecha]
      properties:
        horaMin:
          type: integer
          description: Hora y minuto del dato (formato HHMM)
          minimum: 0
          maximum: 2359
          example: 1430
        tempMedia:
          type: number
          description: Temperatura media del aire (ºC)
          example: 14.0
        humedadMedia:
          type: number
          description: Humedad relativa media (%)
          example: 80.4
        velViento:
          type: number
          description: Velocidad media del viento (km/h o m/s segun config)
          example: 20.0
        precipitacion:
          type: number
          description: Precipitacion acumulada (mm)
          example: 200.2
        estacion:
          type:
            - 'null' 
            - string
          description: Identificación de la estación asociada
          example: "CC01"
        estaciones:
          type: array
          items:
            type: 
              - 'null'
              - string
            example: "CC01, CC02"
        provincia:
          type: 
            - 'null'
            - string
          description: Identificación de la provincia asociada
          example: "CC"
        fecha:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00"
    DiaDTO:
      type: object
      required: [tempMedia, tempMax, tempMin, humedadMedia, humedadMax, humedadMin, velViento, precipitacion, etpMon, pepMon, estacion, provincia, fecha]
      properties:
        tempMedia:
          type: number
          description: Temperatura media del aire (ºC)
          example: 14.0
        tempMax:
          type: number
          description: Temperatura máxima diaria (ºC)
          example: 21.0
        horMinTempMax: 
          type: integer
          description: Hora y minuto de la temperatura maxima (formato HHMM)
          minimum: 0
          maximum: 2359
          example: 1430
        tempMin:
          type: number
          description: Temperatura minima diaria (ºC)
          example: 11.0
        horMinTempMin:
          type: integer
          description: Hora y minuto de la temperatura minima (formato HHMM)
          minimum: 0
          maximum: 2359
          example: 1120
        humedadMedia:
          type: number
          description: Humedad media diaria
          example: 11.9
        humedadMax:
          type: number
          description: Humedad maxima diaria
          example: 20.2
        horMinHumMax:
          type: integer
          description: Hora y minuto de la humedad maxima (formato HHMM)
          example: 2000
        humedadMin:
          type: number
          description: Humedad minima diaria
          example: 11.0
        horMinHumMin:
          type: integer
          description: Hora y minuto de la humedad minima (formato HHMM)
          example: 1100
        velViento:
          type: number
          description: Velocidad media diaria del viento
          example: 11.2
        velVientoMax:
          type: number
          description: Velocidad máxima diaria del viento (km/h o m/s)
          example: 20.4
        precipitacion:
          type: integer
          description: Precipitacion diaria total (mm)
          example: 14.2
        etpMon:
          type: number
          description: Evapotranspiración por Penman-Montheith
          example: 1.12
        pepMon:
          type: number
          description: Precipitación efectiva por Penman-Montheith
          example: 2.3
        estacion:
          type:
            - 'null' 
            - string
          description: Identificación de la estación asociada
          example: "CC01"
        estaciones:
          type: array
          items:
            type: 
              - 'null'
              - string
            example: "CC01, CC02"
        provincia:
          type: 
            - 'null'
            - string
          description: Identificación de la provincia asociada
          example: "CC"
        fecha:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00"
    SemanaDTO:
      type: object
      required: [anio, semana, tempMedia, tempMax, tempMin, diaHoraTempMax, diaHoraTempMin, humedadMedia, humedadMax, humedadMin, diaHoraHumMax, diaHoraHumMin, velViento, precipitacion, etpMon, pepMon, estacion, provincia]
      properties:
        anio:
          type: integer
          description: Año natural del dato
          example: 2025
        semana:
          type: integer
          description: Número de semana (1-52/53)
          minimum: 1
          maximum: 53
          example: 22
        tempMedia:
          type: number
          description: Temperatura media semanal
          example: 11.1
        tempMax:
          type: number
          description: Temperatura maxima semanal (ºC)
          example: 22.1
        tempMin:
          type: number
          description: Temperatura minima semanal (ºC)
          example: 10.0
        diaHoraTempMax:
          type: string
          format: date-time
          description: Dia y hora en la que se registró la temperatura maxima (formato "DDHHMM")
          example: "121100"
        diaHoraTempMin:
          type: string
          format: date-time
          description: Dia y hora en la que se registró la temperatura minima (formato "DDHHMM")
          example: "211100"
        humedadMedia:
          type: number
          description: Humedad media semanal
          example: 10.0
        humedadMax:
          type: number
          description: Humedad maxima semanal 
          example: 20.0
        humedadMin:
          type: number
          description: Humedad minima semanal
          example: 10.0
        diaHoraHumMax:
          type: string
          format: date-time
          description: Dia y hora en la que se registró la humedad máxima (formato "DDHHMM")
          example: "122200"
        diaHoraHumMin:
          type: string
          format: date-time
          description: Dia y hora en la que se registró la humedad mínima (formato "DDHHMM")
          example: "210000"
        velViento:
          type: number
          description: Velocidad media del viento semanal
          example: 11.0
        velVientoMax:
          type: number
          description: Velocidad máxima del viento semanal
          example: 20.0
        precipitacion: 
          type: number
          description: Precipitacion acumulada semanal (mm)
          example: 12.0
        etpMon:
          type: number
          description: Evapotranspiración por Penman-Montheith
          example: 1.12
        pepMon:
          type: number
          description: Precipitación efectiva por Penman-Montheith
          example: 2.3
        estacion:
          type:
            - 'null' 
            - string
          description: Identificación de la estación asociada
          example: "CC01"
        estaciones:
          type: array
          items:
            type: 
              - 'null'
              - string
            example: "CC01, CC02"
        provincia:
          type: 
            - 'null'
            - string
          description: Identificación de la provincia asociada
          example: "CC"
    MesDTO:
      type: object
      required: [anio, mes, numDias, tempMedia, tempMax, diaHoraMinTempMax, tempMin, diaHoraMinTempMin, humedadMedia, humedadMax, diaHoraHumMax, humedadMin, diaHoraHumMin, velViento, precipitacion, etpMon, pepMon, estacion, provincia]
      properties:
        anio:
          type: integer
          description: Año anual del dato
          example: 2025
        mes:
          type: integer
          description: Número del mes (1 a 12)
          minimun: 1
          maximum: 12
          example: 8
        numDias:
          type: integer
          description: Numero de dia en el mes
          example: 21
        tempMedia: 
          type: number
          description: Temperatura media mensual
          example: 23.0
        tempMax:
          type: number
          description: Temperatura máxima mensual (ºC)
          example: 25.0
        diaHoraMinTempMax:
          type: string
          format: date-time
          description: Día y hora de la temperatura maxima
          example: "2025-01-01T00:00:00"
        tempMin:
          type: number
          description: Temperatura minima mensual (ºC)
          example: 21.0
        diaHoraMinTempMin:
          type: string
          format: date-time
          description: Día y hora de la temperatura minima
          example: "2025-01-01T00:00:00"
        humedadMedia:
          type: number
          description: Humedad media mensual
          example: 11.1
        humedadMax:
          type: number
          description: Humedad máxima mensual
          example: 14.0
        diaHoraHumMax:
          type: string
          format: date-time
          description: Día y hora de la humedad máxima
          example: "2025-01-01T00:00:00"
        humedadMin:
          type: number
          description: Humedad mínima mensual
          example: 8.0
        diaHoraHumMin:
          type: string
          format: date-time
          description: Día y hora de la humedad mínima
          example: "2025-01-01T00:00:00"
        velViento:
          type: number
          description: Velocidad media mensual del aire
          example: 11.2
        precipitacion:
          type: number
          description: Precipitacion total mensual (mm)
          example: 40.0
        etpMon:
          type: number
          description: Evapotranspiración por Penman-Montheith
          example: 1.12
        pepMon:
          type: number
          description: Precipitación efectiva por Penman-Montheith
          example: 2.3
        estacion:
          type:
            - 'null' 
            - string
          description: Identificación de la estación asociada
          example: "CC01"
        estaciones:
          type: array
          items:
            type: 
              - 'null'
              - string
            example: "CC01, CC02"
        provincia:
          type: 
            - 'null'
            - string
          description: Identificación de la provincia asociada
          example: "CC"